+++ canu/functions/__init__.py
@@ -1,0 +1,2 @@
+from .retrieve_from_web import retrieve_from_web, retrieve_from_web_json
+from .generate_image import generate_image, generate_image_json