+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## 0.3.0 (in development)
+## 0.3.0 (2024-06-04)
 * Add the `canu.functions.retrieve_from_web` method.
 * Add the `canu.functions.generate_image` method.
 * Add the `canu.EventHandler.on_image_file_done` method.